Add custom auth errors for responding to authentication information requests
Summary: When replying to authentication information requests (AIRs) with error messages, the HSS should not just send diam.UnableToComply. Instead, it needs to send the error code specified in 3GPP TS 29.272 section 5.2.3.1.3. This change adds two new kinds of custom auth errors, so that when the HSS needs to pick which error message to send, it has sufficient information to send the correct error code.

Reviewed By: fishlinghu

Differential Revision: D13994502

fbshipit-source-id: c33ae68fcd283fe4a6da5ec17c0f574ddfaea968