Use custom subscriber storage errors to reply with different diameter error codes
Summary: Previously, if the HSS failed to access subscriber data, it would always return an error code indicating that the subscriber was unknown. This changes the HSS to check that the error is an UnknownSubscriberError before return an unknown subscriber error. If the error isn't an UnknownSubscriberError, then the HSS now outputs a transient error indicating that the request can be retried at a later time.

Reviewed By: fishlinghu

Differential Revision: D13995127

fbshipit-source-id: 3a85395e2dfc003b0208027c1a2fad4bb6dfb78e