Ignore empty flow stats reply in enforcement_stats (#161)
Summary:
Pull Request resolved: https://github.com/facebookexternal/magma/pull/161

Previously, when `enforcement_stats` receives stat reply with no stats in it, it would reset current usage, resulting in reporting the total usage on the next poll, instead of the delta usage.

This issue occurs when other apps (such as `meter_stats` when there are no subscribers, testing app) sends a stats request that would result in an empty reply.

This diff fixes the issue by ignoring the stats reply if there are no stat inside, since there is nothing to reply anyways.

Reviewed By: xjtian

Differential Revision: D14158802

fbshipit-source-id: c464bbb6ee7524cfa1831f0b644177b05d63d250