Refactor message handling to remove code duplication
Summary: When the HSS receives a diameter message, it delegates the handling of it to a diam.HandlerFunc. The code for these handlers was identical, except that a different function was being invoked to construct the response. This factors out the code so that the duplication is removed which will make it easier to implement handlers for more types of messages.

Reviewed By: mpgermano

Differential Revision: D14071823

fbshipit-source-id: 1a70d19831489f4b8a562aa27045a48e14dcff92