Update imsi as required in deactivate flow comment in enforcement
Summary:
This diff updates the comment and arg check logic in deactivate flow to ensure imsi is none.

I tested the function using pipelined cli and if imsi is none, the function would break instead of properly deleting the rule for all subscribers. To properly support the deletion, we would need to update the qos rate limiting to allow deletion of all flows of a rule. The function is currently only used by session manager and there is no use cases for deactivating a rule for all subscirbers, so we can just mention imsi is required for now.

Reviewed By: fishlinghu

Differential Revision: D14087257

fbshipit-source-id: 75fadc5fa2eb97c78e719f63e1168b0a53cd6e91