Drop all non-http traffic when a captive portal is enabled.
Summary:
Openflow doesn't support a "not match" semantic to enable if/else style
branching. Several solutions can be used to enable this behavior, including
using a secondary scratch table with 2 flows: one that matches, and the other
one for the else case.

Here, we can use this model to always drop non-http traffic when a captive
portal is enabled by making the else case a drop action.

Reviewed By: rpraveen

Differential Revision: D13508884

fbshipit-source-id: a883702661a48ab6344a01a8abd84cac801ec7e2