Convert auth scheme to a string before storing it in MARs
Summary: The auth scheme was cast directly from an int to a string when a multimedia authentication request (MAR) was proxied from gRPC to diameter. This caused the HSS to see invalid auth schemes in MARs. Instead, now the auth scheme is converted to a string constant before it is stored in the MAR.

Reviewed By: xjtian

Differential Revision: D14216213

fbshipit-source-id: 7f78ea21cf1aed3cd1841e5529c3c620f53d1260