Stop incrementing IND when generating auth vectors (#10)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/10

Per 3GPP TS 33.102 Appendix C3.4, if a new request comes from the same serving node as the previous request (which will always be the case since we use a FeG), then the index value used for the new request should be the same as the previous request. So, instead of allocating the value cyclically within its range, it should instead be kept as a constant.

Reviewed By: fishlinghu

Differential Revision: D14226835

fbshipit-source-id: 8b5b9da516485596300696196e9d615fbfa23af9