Fix FeG mconfig init refresh error
Summary:
Currently, when the mconfig init function is called for the Federated Gateway, there's an error
that the mconfig for a given service is not found. This stems from the ocr8r mconfig protos not having
been registered before the mconfig init() function is run. This produces an error such as `magma.mconfig.ControlProxy` not found.

This diff adds the orc8r mconfig protos to the import list to ensure that these message types are known before the function triggers.
Since the protos don't have to be used, they are prefixed with `_`

Reviewed By: vikg-fb

Differential Revision: D14236492

fbshipit-source-id: 67fe5135b99c9af0c666ef7f90b7c1ced9d6aefd