Update enforcement stats to use rule match flows
Summary: This diff updates `enforcement` and `enforcement_stats` to use a new scratch table for usage reporting. Previously, we are reusing the flow match flows used to match rule id to report usage. With this diff, `enforcement` will forward traffic to a scratch table and only install rule match flows to reduce the number of flows required.

Reviewed By: amarpad

Differential Revision: D14019380

fbshipit-source-id: 081b1c9ce788d0a80502c6da0567712807497177