Restructure how unexpected TR069 messages are handled
Summary:
The state transitions for each eNB device do not take into account that an eNB will send an Inform message at a regular interval even if the device is in the provisioning process. To handle this scenario more gracefully, EnodebAcsState will raise an UnhandledTr069MessageError when an unexpected message type is received, and the state machine will transition to the appropriate state before handling the incoming message.

The state machine will now gracefully handle unexpected Inform and Fault messages. For unexpected Fault messages, the state machine will transition to an error state, instead of flooding logs with repeated exceptions.

Reviewed By: xjtian

Differential Revision: D14211153

fbshipit-source-id: dbc3f40fd4531a06b38e88b6cf432edf610c5743