parse networkID, gatwayID from cloud metric names
Summary: Metrics that are submitted from "ConsumeCloudMetrics" are given the networkID "cloud". This makes it impossible to expose to the correct network since we label based on the NetworkID. Most (if not all?) of the metrics I've seen have this formatted name of <metricName>_gatewayID_<gatewayID>_networkID_<networkID>. Parsing those out in the prometheus exporter to apply the correct labels will allow us to expose these cloud metrics.

Reviewed By: xjtian

Differential Revision: D14187778

fbshipit-source-id: ca180eddf33cba5e5a4c773bb459eab80dfdbac7