Fix FeG obsidian handler gw config type
Summary:
This diff corrects a bug in which the obisidian handler for the federated gateway config type
was set as the network config. This resulted in the feg mconfig always being composed of the network
config type, even if the gateway type was set in swagger. This diff fixes this bug to ensure mconfig
works properly.

Reviewed By: emakeev

Differential Revision: D14304109

fbshipit-source-id: 847403fb56b3c34dde9f746be9ca77a539c61b8c