ansible role now reads service_registry from remote /etc/configs/ not from host
Summary:
Since prod does not execute modules when provisioning, the service_registry variable (defined in modules) was never defined. This led to proxy failing because proxy relies on that variable.

I created a new role, service_reigstry, so that it can be included when needed.
I made it so that it'll look up in /etc/magma/configs to see which modules are there and it'll use that information to fetch the service_registry files.

Reviewed By: xjtian

Differential Revision: D14304843

fbshipit-source-id: 47d3cf8fbf6b8034c128a8ae38e7c4264b35a046