ansible role now reads service_registry from remote /etc/configs/ not from host (#29)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/29

Since prod does not execute modules when provisioning, the service_registry variable (defined in modules) was never defined. This led to proxy failing because proxy relies on that variable.

I created a new role, service_reigstry, so that it can be included when needed.
I made it so that it'll look up in /etc/magma/configs to see which modules are there and it'll use that information to fetch the service_registry files.

Reviewed By: xjtian

Differential Revision: D14304843

fbshipit-source-id: 890a01c0d459765d66c3772417eb9752c5d173c6