Add meter flow deletion scheduling to avoid underreporting usage
Summary:
**Background**:
Metering in PipelineD is implemented using 3 controllers: `meter`, `meter_stats`, and `subscriber`.
- `meter` is responsible for installing flows for each subscriber.
- `meter_stats` can be configured to poll stats on the flows installed by `meter` and report them to `meteringd`.
- `subscriber` can be configured to poll the list of subscribers from `mobilityd` and remove metering flows for subscribers that are deleted.

Previously, `subscriber` deletes the flows for deleted subscriber right away, resulting in underreporting of usage stats for the deleted subscriber.

This diff fixes the underreporting issue by delaying the flow deletion until a stats reply event for the metering table is received, which is also used by `meter_stats` to report usage.

Stats reply events are chosen to coordinate flow deletion as this approach does not require introducing additional dependency between `meter_stats` and `subscriber` compared to other approaches such as adding an interface in `subscriber` that `meter_stats` can call to signal that the using is reported. This approach does not guarantee that all usage stats are reported all the time, since `meter_stats` can technically fail to process the stats reply. I think it is acceptable to ensure deletion of flows at the cost of underreporting the usage in these degenerate scenarios.

Note that if polling is off, then the deletion process is unchanged.

Reviewed By: amarpad, xjtian

Differential Revision: D14199932

fbshipit-source-id: 642df993e32dff4e187281f4e24a9151ee212f05