Update rpc servicer for policy apps change
Summary:
This diff updates rpc servicer to call both `enforcement` and `enforcement_stats` when a rule is activated/deactivated, since `enforcement_stats` manages its own flows now.

Failure handling logic is also added to ensure that if the activation fails in one controller, the operation will be reverted in the other controller so we don't end up in inconsistent state. Success will only be reported if flows are installed in both controllers.

Reviewed By: amarpad, xjtian

Differential Revision: D14098435

fbshipit-source-id: 4f639a025d59b536c797883b59fd76b1e3e46da0