Add a command in pipelined cli to get table assignments
Summary:
With the introduction of service chaining, table assignment is now dynamic and we can no longer simply refer to table X for flows for an app. This makes debugging hard, as the table assignment needs to be figured out from studying the flows in the flow dump.

This diff introduces a new command in `pipelined_cli`, `table_assignment`, which can be used to print a table of table assignments of each app. An optional argument `--apps` can be used to only filter the app assignment of interest.

A method is introduced in service manager to expose the app to table mapping and a new rpc endpoint is added to pipelined to expose this information. We expose the assignment with apps instead of services since this command is meant for internal debugging, and it is extra work to get the app -> service reverse mapping.

Reviewed By: amarpad

Differential Revision: D14402663

fbshipit-source-id: 2c9e4ab73178fb9a4015bb9096e088e1b504b80a