Fix use-after-free bug in partial reset handling
Summary:
We are having an issue with MME involving partial reset. The incoming initialuemessage packet is processed but fails before it sends a response to eNB. This causes the eNB to timeout and send a partial reset message (slightly strange behavior) for the involved UE. This hits a use after free bug present in partial reset.

This diff changes the ue_to_reset_list to use raw ints for enb/mme_ue_s1ap_id's instead of pointers to memory with unknown lifetime -- preventing the possibility of uaf.

Reviewed By: anooptomar29

Differential Revision: D14388597

fbshipit-source-id: 89944780738df9e2137453e5656138d3b0b38a40