Create SWx CLI for debugging and testing (#55)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/55

With the recent creation of the SWx service, we now have a service that can translate gRPC into
SWx protocol over diameter. For upcoming integration and deployments, having a tool to easily send requests
to the service will be useful both for verifying service correctness and connectivity with the HSS.

This CLI's main functionality is to send MAR diameter message to the HSS and receive back an MAA. This can be
done either through an existing SWx service that is running, or using a built-in proxy service that is a part
of this CLI. This allows us to test both if our local service is configured correctly  as well as test HSS connectivity
before this service is deployed.

This CLI achieves its functionality either with the use of flags or by running in 'interactive mode' that
allows necessary parameters to be entered one by one for a more user friendly experience.

Reviewed By: vikg-fb

Differential Revision: D14380381

fbshipit-source-id: f39be7a8f6185143b24142dcaa2184560b3fd1c4