Convert subscriber auth key and opc between hex and base64 (#61)
Summary:
This fixes issue #59. The auth key and opc are converted to base64 when they are sent to the backend. Then, when editing subscriber data, it is converted back to hex so it can be presented to the user. A followup change to the backend will be needed to validate that the auth key and opc are the correct size.
Pull Request resolved: https://github.com/facebookincubator/magma/pull/61

Reviewed By: rckclmbr

Differential Revision: D14433104

Pulled By: vikg-fb

fbshipit-source-id: cc0da7f044690250c8fa1eedae3d6c9b0e9ccd1d