Properly delete all enforcement_stats flows on subscriber detach (#64)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/64

Currently, the `deactivate_flow` rpc endpoint deleted enforcement_stats flows for a given subscriber + ruld_ids combination by incrementing the current version of these flows, which schedules a deletion. However, it did not handle the other use case of deleting all flows for a subscriber if rule_ids is empty. As a result, these enforcement_stats flows are not deleted, and enforcement_stats continues reporting stats to sessiond, resulting in errors in sessiond's side.

This diff adds a check in the rpc servicer for empty rule_ids. If it is empty, then the version of all flows for the subscriber will be incremented so they will be deleted.

`pipelined_cli` is also updated to test this case. When --rule_ids is not specified, an empty list will be sent to delete all flows, rather than a default list of rules.

Reviewed By: sciencemanx

Differential Revision: D14459830

fbshipit-source-id: d93922a7be69e2ab7819bb1d79d2ce9b2f8e5a1b