Implement generic command framework in gateway (#73)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/73

This diff implements the generic command framework on the gateway side.

When `magmad` starts up, it now looks for the following config in `magmad.yml`. If it exists, the `GenericCommand` RPC call will succeed, otherwise it will return an error.
```
generic_command_config:
  # location of command executor module
  module: ...
  class: ...
```
It looks up the class located in that module and sets it as the command executor. The command executor stores a dictionary that maps names (as strings) to functions to execute. Using the shell command executor, it reads a list of shell commands from the config into its dictionary, creating a subprocess per shell command in the list when ran, as shown below. However, users can choose to create their own modules to extend this functionality (such as running native python).
```
generic_command_config:
  module: magma.magmad.generic_command.shell_command_executor
  class: ShellCommandExecutor
  shell_commands:
    - name: run_something
      command: "/path/to/script/ foo"
    - name: do_something
      command: "/another/command {}"
      allow_params: True
```
The shell command executor is the default command executor. It returns the return code, stdout, and stderr after the command runs. In addition, parameters may be specified if `allow_params` is set to true for that command. It will read the list `param_list` from the parameters dictionary and format the command string with the list of parameters.

When a client calls `GenericCommand`, the servicer in `magmad/rpc_servicer.py` will run the command located in the command executor's table and return the result of that command. If more than `timeout_secs` (default 15) seconds pass while running the command, it is considered a timeout and an error is returned.

Reviewed By: rpraveen

Differential Revision: D14356008

fbshipit-source-id: d8be129edfb3e54cc9b41871ed04f282240ab73a