Create new config_cli for gateway
Summary:
This diff adds a new config_cli which allows users to modify the config_level by changing the override YAML config. This will make it easier to change the log level for a particular service.

Note that because the cli will load the override YAML file and write the object in memory back as a file, any comments in the override file will be lost.

Differential Revision: D14490519

fbshipit-source-id: ea070495a04425db3bfa01a073b2327e33789f55