LocationsMap - show locations by type on a map
Summary:
Created a new page for the map which will show all the locations by their type with different colored dots.
If no selection is made we should only bring a subset of locations, regardless of their type. We still don't support the "first" filter so this is not implemented yet.

Reviewed By: noamsch

Differential Revision: D14439275

fbshipit-source-id: d7b73ae64fe3e46447e7aebe479f1684b16e31e4