Validate auth key and opc length (#62)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/62

This is a fix for [issue #59](https://github.com/facebookincubator/magma/issues/59). The backend was not validating the length of the auth key and opc. This allowed invalid keys to get to SubscriberDB running in the gateway. This change adds validation so that invalid subscriber data cannot be added to the database using NMS.

Reviewed By: sciencemanx

Differential Revision: D14430821

fbshipit-source-id: 76e22cf157c44beddbbf07c4d1ee5361445973a3