Switch to YAML configs for log level
Summary:
This diff changes all python service303 services to use the local YAML config for log level, instead of mconfig. This allows a way for log level config to be persisted locally through an override config file.

A new `log_level` property is added to the config for each service for clarity. This is not needed since in the event that no `log_level` is found in the comment, the service will default to INFO.

Reviewed By: xjtian

Differential Revision: D14490520

fbshipit-source-id: caaac5d24804bfd39d9b9d9a4fc4b3c7d823a47b