Create new config_cli for gateway
Summary:
This diff adds a new config_cli which allows users to modify the log level by changing the override YAML config. This will make it easier to change the log level for a particular service. It also gives a command to get the current log level, since we log level can be defined in the default config and overridden by the override config.

Note that because the cli will load the override YAML file and write the object in memory back as a file, any comments in the override file will be lost.

Reviewed By: xjtian

Differential Revision: D14490519

fbshipit-source-id: 721edad456fd12547f0982018c1c30d79b15460a