Add tests for ReloadServiceConfig in MagmaService (#79)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/79

Ian is working on a command line tool to adjust service configurations
without requiring a service restart. To support this we need a way for
each service to accept these requests.

This diff adds unit tests for ReloadServiceConfig in MagmaService.cpp.
Specifically, it tests that by default MagmaService::ReloadServiceConfig
returns an unsupported result code. Then, sets the callback to a
succeeding and failing config change and tests that it receives the
correct return codes (success/failure). Finally, the callback is cleared
and we check that it again returns an unsupported result code.

Reviewed By: ian952

Differential Revision: D14482020

fbshipit-source-id: 9b442430e34c7e039cd88d9ae31f0174e42999d6