Add reload_config_callback to service.py for ReloadServiceConfig (#77)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/77

Ian is working on a command line tool to adjust service configurations
without requiring a service restart. To support this we need a way for
each service to accept these requests.

This diff adds support for registering a custom config reload callback
from python services which use MagmaService. This callback returns a
bool indicating if the reload succeed. If the service does not support
reloading their config, they should not register a callback --
MagmaService will automatically return an unsupported result for GRPC
calls to the service.

Reviewed By: ian952

Differential Revision: D14482023

fbshipit-source-id: 95ccc06f8f806efb48aca9b68927e90cb5531f5f