Add ReloadConfigResult service303 RPC method and stubs (#76)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/76

Ian is working on a command line tool to adjust service configurations
without requiring a service restart. To support this we need a way for
each service to accept these requests.

This diff adds a GRPC method "ReloadConfigResult" that will handle be
the endpoint for the cli. Here, only stubs (return unsupported) are
added to implement the RPC method. In a future diff, callbacks will be
added to allow each service to customize its reload service logic.

Reviewed By: xjtian, ian952

Differential Revision: D14482021

fbshipit-source-id: 622c0fb280a65ae28250b730e33e256682f39750