Add feature flag for using the EPS Authentication service
Summary:
This change adds a feature flag which will allow using the EPS authentication service instead of local subscriber DB or the s6a proxy. Currently, the flag must be manually edited in the `mme.yml` file, but in a future change this will be extended so that it can be set using swagger (like the relay_enabled flag). The logic to decide which service handles authentication requests is as follows:

1) If the relay enabled flag is true, then the S6a Proxy in the FeG will receive the requests
2) Otherwise, if cloud_subscriberdb_enabled is true, then the EPS authentication service will receive the requests
3) If neither flag is true, then local subscriber db is used

This feature flag is designed so that when it is false, nothing is changed. This will make deployment of the eps_authentication service easier. After deployment, we plan to cleanup the flags since it is not needed to have two bool flags like this.

Reviewed By: fishlinghu

Differential Revision: D14448120

fbshipit-source-id: 7c8dc4c3713ab1df98bf06b3e577000acc2206cb