Fix flaky test: TestReadSuccessAndFailWithTimeout
Summary: TestReadSuccessAndFailWithTimeout failed in CI since the timeout was not long enough. This allowed a goroutine to read from the safe buffer when it should not have been able to. To prevent this test from failing in the future, go channels can be used for signalling instead of waiting for a fixed amount of time.

Reviewed By: fishlinghu

Differential Revision: D14543306

fbshipit-source-id: 357c5b89c7152bc7175af9fef1ca0c5c25f68180