Add missing AVPs to AIAs and ULAs (#105)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/105

This change adds AVPs to the HSS lite that a real HSS would send in authentication information answers (AIAs) and update location answers (ULAs). That way, the HSS lite behaves more similarly to a real HSS and is a more effective testing component.

The following AVPs were added:
* Auth session state
* Vendor specific application id
* E-UTRAN vector item number

Also, the experimental result AVP is no longer added on success.

Differential Revision: D14544002

fbshipit-source-id: 2de23fd0819a94d4e91b00a7962e6ce16f7d84ab