Differentiate between permanent and transient storage errors
Summary:
When the EPS Authentication service fails to lookup a subscriber, then either the subscriber is not in the database, or there was a transient storage error. This changes the `lookupSubscriber` function to differentiate between these two kinds of errors as specified by 3GPP TS 29.272. A transient error indicates to the caller that they can retry the request at a later time whereas a permanent error signals that the request should not be retried.

Also, while adding a unit test, I noticed that in `ul_test.go` there is a typo where the message variable name was `air` instead of `ulr`. I corrected the typo in this change as well.

Differential Revision: D14548337

fbshipit-source-id: df7089c4170ee061eb3d7c32d93d5931ef90428c