Create federated gateway README
Summary:
The FeG currently does not have any documentation. This README serves to add basic sections
about the purpose of the FeG, as well as basic steps on how to set it up, upgrades, installations, and
an FAQ. This will hopefully be useful to both the Magma team and 3rd party developers.

Reviewed By: sciencemanx

Differential Revision: D14473850

fbshipit-source-id: 088cafa37089fc3c91ed2296bc7e63e153b89a66