Change PolicyDB to use glog for logging
Summary:
Because we don't configure PolicyDB to log with glog, its logging messages get printed to stdout. In sessiond, there is a thread to repeatedly sync rules in PolicyDB, which repeatedky prints a message in the log that cannot be disabled.

This diff fixes the issue by compiling policydb with `LOG_WITH_GLOG` flag, so the message can be filtered by verbosity.

Reviewed By: sciencemanx

Differential Revision: D14506421

fbshipit-source-id: d7cf3bff6165750edd79f52c85ae0a2c76bfac63