Add pre-configured subscribers on startup to testcore HSS
Summary:
Recent testing of the HSS has shown a need to be able to configure subscribers in
the test HSS on start-up rather than using the hss_cli. Preconfigured subscribers allow a
repeatable test environment with less variables in play. This diff serves to add subscribers
that are listed in an `hss.yml` file as the HSS service begins running. These subscribers are
given a basic profile.

If the need arises in the future to allow for more configurable options, this can easily be added.
While helpful, the best long term solution would be to stream subscribers from the orc8r, so that
subscribers are stored state, rather than being added on startup.

Reviewed By: vikg-fb

Differential Revision: D14573510

fbshipit-source-id: c333f1131af5e6b0335e17b49b65ee5c5531118f