Add remote FeG connection for swx_proxy (#118)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/118

This diff serves to add the ability to connect a federated gateway VM to different cloud instances. This serves
to allow the local feg vm to be tested against a running cloud instance. Additionally, a remote client is added to
the swx_proxy service to allow for the use case of the eap service and swx_proxy service running on different machines.

Reviewed By: fishlinghu

Differential Revision: D14536789

fbshipit-source-id: 013d21dde9949ee26b231440a0609e38bd6ff9b6