Fix cloud build: change from RegisterConfigManager to RegisterSerdes
Summary: A recent change changed a function name from RegisterConfigManager to RegisterSerdes. This change was not applied to the EPS authentication service (since the service is new and was added after the function rename, but before the rename was landed). This fixes the rebase conflict and allows precommit to succeed again.

Reviewed By: ian952

Differential Revision: D14599438

fbshipit-source-id: d2be02cbf0d509f7a93dab7cb2963a271942ea3a