Implement config reloading for python service303 services
Summary:
This diff enables python service303 services to reload the local config. When the ReloadServiceConfig rpc endpoint is called, the service will reload the config and configure logging based on the new config. This allows logging level config to be applied dynamically without a service restart.

The previous reload callback is removed, as this is not going to be useful for the time being, since log level reloading is handled by service303.

Reviewed By: amarpad, xjtian

Differential Revision: D14540765

fbshipit-source-id: 84e43261949a777a4f12b10673bb1ca425c4fb65