Add config reload rpc call in config_cli
Summary:
This diff updates config_cli to make an rpc call to the service to reload the config when setting the log level.

For services that support reloading, this will allow the log level to be updated dynamically when changing the override while changing override config. Running the `set_log_level` command will change the log level for the current running instance, as well as for future instances.

Reviewed By: amarpad

Differential Revision: D14517262

fbshipit-source-id: 8b10a3d916f047e328ff2810300fc28234c54cb5