Change implementation of length validation for auth key and opc
Summary: The auth key, "DtR1RRaOr+LDnAdYKae2Hw==" was rejected by swagger as having an incorrect length even though it is the correct length (16 bytes). So, swagger does not seem to be a reliable way of validating the length of base64 encoded strings. This change removes the length specification from the swagger definition and instead checks the length after the swagger validation.

Reviewed By: fishlinghu

Differential Revision: D14570502

fbshipit-source-id: bd16b0971f38e91b9b1150516b1a08ea3dd9d287