Add basic docusaurus website
Summary: Adds a script which can generate a docusaurus docs website. Documentation will be stored in `docs/readmes`, docusaurus files are stored in `docs/docusaurus` and the website generated files are stored in `docs/web`. `./docusaurus/create_docusaurus_website.sh` can be run to generate the files in the `docs/web` folder needed to serve the website. The website files are built and run inside of a docker container to ensure reproducibility.

Reviewed By: rpraveen

Differential Revision: D14484858

fbshipit-source-id: 037fa1c5986d95e7580ca7291c992cc5a4c4c088