Remove unused functions that cross-task call into s1ap
Summary:
Currently, there is call path from NAS to MME_APP that manipulates S1AP state. This poses a threat to stateless S1AP/MME as it breaks assumptions ITTI makes. Luckily, this function is never called so we can safely delete it.

This diff removes the offending call flow.

Reviewed By: ssanadhya

Differential Revision: D14565850

fbshipit-source-id: cc68db8408d07d4229b9b56b586d2cd929a99315