Include C/C++ unit tests during CI run
Summary:
Currently, C/C++ unit tests are not being run on CI, because it calls `make test_all` in the python directory, instead of `make test` in the outer directory. This means that tests for OAI, and sessiond are not being run.

`make test_oai` is also failing to build due to include error with service303 and service registry. To fix build for now, these tests are removed from the makefile.

Reviewed By: ssanadhya

Differential Revision: D14609060

fbshipit-source-id: f942e3c352d077b1c1aaea40cac099dd71093353