Reboot request will send after responding to Inform with InformResponse
Summary:
While eNB will be able to respond to an empty http response when an Inform is sent to the AutoConfigurationServer (hosted by enodebd), it will not be able to respond to a Reboot. To handle this, enodebd will now make sure an InformResponse is sent back to the eNB before a Reboot request is sent.

**eNB -> ACS**
  Inform

**ACS -> eNB**
  InformResponse

**eNB -> ACS**
  (empty HTTP request)

**ACS -> eNB**
  Reboot

**eNB -> ACS**
  RebootResponse

**ACS -> eNB**
  (empty HTTP request)

Reviewed By: sciencemanx

Differential Revision: D14582361

fbshipit-source-id: 82bbd6bc5fef4666d1f5f62ac5dea0d404b5277f