Flow for migration scripts
Summary:
Add flow to migration scripts.

Flow complains that you need to import as a type instead of using require. However, that is syntactically different. I'm pretty sure sequalize does something weird with how it uses the types, but I don't want to investigate further.

Type annotations are stripped out, however when doing the require, we get an object of string -> object. Thus, I will leave the code as is and tell flow to ignore that one error.

Reviewed By: rckclmbr

Differential Revision: D14634347

fbshipit-source-id: f1823cfa57da7d195326cd9db1ae55d4fae8c3b9