Fix HSS-Lite resync sqnMS generation
Summary:
In recent integration testing, it was discovered that the HSS-Lite in testcore doesn't
resynchronize properly. This is a crucial step for the HSS working with real hardware as we don't
store the sequence number for a subscriber past runtime. As such, once the sequence number gets high enough
on the UE, the re-sync sequence would continuously trigger without working. This fix corrects a bug
with the generation of the sqnms. As such the HSS now successfully resyncs with a UE on startup.

Reviewed By: emakeev

Differential Revision: D14639994

fbshipit-source-id: 3c3d84c2d02e009f2734b908f7badc3b53eb5248