Adding Sequence Diagram for Attach Call Flow
Summary:
In order to visualize the attach call flow in Magma, this change adds a sequence
flow diagram. The file Attach_call_flow_in_Magma.txt can be uploaded to
sequencediagram.org to edit and to export the .svg. or .jpg image.
The README explains the color scheme used in the diagram.

Reviewed By: amarpad

Differential Revision: D14662640

fbshipit-source-id: 6b6c59559173eec871e1d7c6f2187d46f220c78f