@flow fbcnms-logging
Summary:
Fix flow for fbcnms-logging.

This means importing the flow-typed files for morgan and winston.
This means fixing the stream write callback since it's incorrect (checked flow types and the actual morgan code).
Adding generic type for winston for flow. This will enable autocomplete when logging.
Ensuring valid log levels are used in the config. (I couldn't find a way to reuse the allowed types without importing the other packages winston uses, and I didn't want to deal with that). Flow will complain if the types don't match, so it should be fine.

Reviewed By: rckclmbr

Differential Revision: D14658533

fbshipit-source-id: 5dddf5e4cce95a48f992e4aa6b7e1316c2f9212c