Initialize subscriber state if it is nil
Summary: Sometimes the subscriber state is nil, which can cause the EPS Authentication service's auth requests to fail. Since resync requests can be used to get the correct sequence number, initializing the sequence number to a default of zero will allow these requests to succeed when the state is nil.

Reviewed By: ian952

Differential Revision: D14673623

fbshipit-source-id: 4d703d4aaeda26829706e469bd6ff2c9ed01bf8d