Fix all eslint errors
Summary:
There are now now more eslint errors, only warnings.
Of the 13 warnings left, only 1 is in inventory, and 1 in nmsx (the rest are in tgnms).

Not sure how to fix the line length in inventory, and the routes.js for magmanms is difficult, so leaving it out for now.

Reviewed By: rckclmbr

Differential Revision: D14670299

fbshipit-source-id: c2c9a6a1a1eb5a42b912839ce11bcd22aae70089