Move cloud_subscriberdb_enabled feature flag from mme.yml to network cellular config (#140)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/140

To make cloud subscriberdb easier to use and test, the feature flag for enabling it is moved from mme.yml to the mconfig system. With this change, the cloud subscriberdb can be enabled from swagger and the flag in mme.yml is removed. Now that the full functionality of the eps authentication service is implemented, this is a next step towards productionizing cloud subscriberdb.

Reviewed By: fishlinghu

Differential Revision: D14668592

fbshipit-source-id: f8fa695d09664c1d07cae7728e5c948c74b62b9c