Report mconfig created at timestamp during checkin
Summary:
This diff adds the created at timestamp of mconfig to the gateway status during checkin.

In order for this to work, magmad is modified to reload the mconfig file on every streamer update, instead of only when the mconfig is changed. This is because the metadata portion of mconfig (which contains created at timestamp) will change on every update, and we always want to send the latest timestamp. Functionally speaking, there should be no change to the mconfig streaming process. Services will only be restarted when their respective config is changed.

Reviewed By: rpraveen

Differential Revision: D14517853

fbshipit-source-id: 74c5e7b569f179de15e7eabf9c5ad848c1418432