Collect mconfig age during checkin
Summary:
This diff updates the gateway status proto to also include mconfig's created at timestamp.

Using the timestamp, we can determine the staleness of mconfig and alert if the mconfig is past a certain threshold. In terms of alerting, a new metric is added for mconfig age, which can be used to set up alerts in ODS.

Reviewed By: amarpad

Differential Revision: D14511808

fbshipit-source-id: ed20c8947bde574f20aec7c44feea8940214d059